requirements.txt
python-dotenv
pytrends
requests
anthropic
beautifulsoup4
